requirements.txt
tweepy
requests
bs4
Pillow
random
